Fix goreleaser config (#27)
There was an error generating the build

```
│ cmd/bot/main.go:48:14: undefined: parseConfig
```

I believe this is because we set the entry point for the application to me `cmd/bot/main.go` causing go to not be able to find the dependencies. THis updates the entry point to be `./cmd/bot`.

There was also a warning for a deprecated flag:

```
 • DEPRECATED:  archives.format  should not be used anymore, check https://goreleaser.com/deprecations#archivesformat for more info
```

This was updated in the `.goreleaser.yaml` file.