Migrate to using cloudcost-exporter metrics (#30)
First attempt at trying to migrate to using cloudcost-exporter metrics. Took a naive approach of adding a config to toggle cloudcost-exporter metrics. Added two new queries for cloudcost-exporter to get the avg cost of
- Cpu cores
- Memory (GiB)

Discovered that cloudcost-exporter does not implement cost of persistent storage(see https://github.com/grafana/cloudcost-exporter/issues/236), so I'll pick that up in a separate PR. 

To test out, follow the setup guide and run the estimator command:

```shell
go run ./cmd/estimator/ \
  -from $PWD/pkg/costmodel/testdata/resource/Deployment.json \
  -to $PWD/pkg/costmodel/testdata/resource/Deployment-more-requests.json \
  -http.config.file ~/.config/dev.yaml \
  -prometheus.address $PROMETHEUS_ADDRESS \
   dev-us-east-0

```

Since this releases the change behind a flag, we can merge this whenever we're comfortable and enable it when we're comfortable. 

- relates #29