Add useragent to headers (#34)
This adds a configuration option for the RoundTripper that sets the useragent to `grafana-kost-estimator`. This helps downstream Prometheus instances understand where the queries that are coming from. If the logs are captured by something such as loki, you can easily filter our queries sent to Prometheus by the kost bot.