Remove all references to OpenCost
Now that the v0.0.5 has been tested out in production, we can
remove all references to OpenCost.

This removes the flags, configuration, and branches in the codebase.
There is still one query that references OpenCost, but that can be
removed in a future patch once cloudcost-exporter emits Azure persistent
volume costs.